Refactor notifications into reusable hook.
